LayoutTests/imported/w3c/web-platform-tests/editing/other/edit-in-textcontrol-immediately-after-hidden.tentative_editor=textarea&hide-target=editor-expected.txt


PASS Wait for load event
PASS execCommand("insertText", false, "typed value") in <textarea>
PASS execCommand("delete") in <textarea>
PASS <textarea> is hidden by "keydown" event listener
PASS <textarea> is hidden by "keypress" event listener
PASS <textarea> is hidden by "beforeinput" event listener
